+++ src/app_types/sync_runable.py
@@ -20,20 +20,18 @@
 # OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
 # OR OTHER DEALINGS IN THE SOFTWARE.
 
-import pytest
-from fakeredis import aioredis
-
-from app_types.fk_log_sink import FkLogSink
-from srv.users.redis_user_state import RedisUserState
-from srv.users.user_step import UserStep
+from pyeo import elegant
 
 
-@pytest.fixture()
-def fake_redis():
-    return aioredis.FakeRedis()
+from typing import Protocol
 
 
-async def test_not_exists_state(fake_redis):
-    got = await RedisUserState(fake_redis, 879435, FkLogSink()).step()
+@elegant
+class SyncRunable(Protocol):
+    """Интерфейс блокирующего запускаемого объекта."""
 
-    assert got == UserStep.nothing
+    def run(self, args: list[str]) -> int:
+        """Запуск.
+
+        :param args: list[str]
+        """
