+++ src/srv/events/event_hook.py
@@ -35,7 +35,7 @@ from pyeo import elegant
 from quranbot_schema_registry import validate_schema
 
 from app_types.logger import LogSink
-from app_types.runable import SyncRunable
+from app_types.sync_runable import SyncRunable
 from settings import Settings
 from srv.events.recieved_event import ReceivedEvent
