+++ src/tests/it/handlers/test_favorites_answer.py
@@ -22,7 +22,7 @@
 
 import ujson
 
-from app_types.logger import FkLogSink
+from app_types.fk_log_sink import FkLogSink
 from app_types.update import FkUpdate
 from handlers.favorites_answer import FavoriteAyatsAnswer
 from integrations.tg.tg_answers import FkAnswer
