+++ src/tests/it/handlers/test_search_city_answer.py
@@ -23,7 +23,7 @@
 import pytest
 import ujson
 
-from app_types.logger import FkLogSink
+from app_types.fk_log_sink import FkLogSink
 from app_types.update import FkUpdate
 from handlers.search_city_answer import SearchCityAnswer
 from integrations.tg.tg_answers import FkAnswer
