+++ src/tests/it/srv/ayats/test_change_favorite_ayat_answer.py
@@ -23,7 +23,7 @@
 import pytest
 import ujson
 
-from app_types.logger import FkLogSink
+from app_types.fk_log_sink import FkLogSink
 from app_types.update import FkUpdate
 from integrations.tg.tg_answers import FkAnswer
 from srv.ayats.change_favorite_ayat_answer import ChangeFavoriteAyatAnswer
