+++ src/tests/it/srv/ayats/test_search_ayat_by_text_callback_answer.py
@@ -23,7 +23,7 @@
 import pytest
 import ujson
 
-from app_types.logger import FkLogSink
+from app_types.fk_log_sink import FkLogSink
 from app_types.update import FkUpdate
 from exceptions.content_exceptions import AyatNotFoundError
 from integrations.tg.chat_id import FkChatId
