+++ src/tests/it/srv/events/test_check_user_status.py
@@ -25,7 +25,7 @@ import pytest
 import ujson
 from eljson.json_doc import JsonDoc
 
-from app_types.logger import FkLogSink
+from app_types.fk_log_sink import FkLogSink
 from integrations.tg.tg_answers import FkAnswer
 from srv.events.check_user_status import CheckUsersStatus
 from srv.events.sink import FkSink
