+++ src/tests/it/srv/prayers/test_prayers.py
@@ -25,7 +25,7 @@ import uuid
 
 import pytest
 
-from app_types.logger import FkLogSink
+from app_types.fk_log_sink import FkLogSink
 from app_types.update import FkUpdate
 from handlers.prayer_time_answer import PrayerTimeAnswer
 from integrations.tg.tg_answers import FkAnswer
