+++ src/tests/it/test_start.py
@@ -24,7 +24,7 @@ import pytest
 import ujson
 from furl import furl
 
-from app_types.logger import FkLogSink
+from app_types.fk_log_sink import FkLogSink
 from app_types.update import FkUpdate
 from handlers.full_start_answer import FullStartAnswer
 from integrations.tg.tg_answers import FkAnswer
