+++ src/tests/unit/integrations/tg/tg_answers/test_answer_fork.py
@@ -22,7 +22,7 @@
 
 import pytest
 
-from app_types.logger import FkLogSink
+from app_types.fk_log_sink import FkLogSink
 from app_types.update import FkUpdate
 from exceptions.internal_exceptions import NotProcessableUpdateError
 from integrations.tg.tg_answers import FkAnswer, TgAnswerFork
