+++ src/tests/unit/srv/prayers/test_invite_set_city_answer.py
@@ -25,7 +25,7 @@ from typing import override
 import httpx
 import ujson
 
-from app_types.logger import FkLogSink
+from app_types.fk_log_sink import FkLogSink
 from app_types.update import FkUpdate, Update
 from exceptions.content_exceptions import UserHasNotCityIdError
 from integrations.tg.tg_answers import FkAnswer, TgAnswer
